backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.2",
    "fastapi>=0.115.5",
    "loguru>=0.7.2",
    "openai>=1.54.4",
    "pillow>=11.0.0",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.17",
    "uvicorn>=0.32.0",
]

